+++ index.html
@@ -11,7 +11,7 @@
     <script src="dist/advancedTypes.js" defer></script>
     <script src="dist/generics.js" defer></script>
     <script src="dist/decorators.js" defer></script> -->
-    <script src="dist/app.js" defer></script>
+    <script type="module" src="dist/bundle.js"></script>
   </head>
   <body>
     <template id="project-input">
