+++ src/examples/interfaces.ts
