+++ src/examples/thirdparty.ts
@@ -1,0 +1,5 @@
+import _ from 'lodash';
+
+declare var GLOBAL: any;
+
+console.log(_.shuffle([1, 2, 3]));
