+++ webpack.config.js
@@ -1,0 +1,28 @@
+/* eslint-disable no-undef */
+const path = require('path');
+
+module.exports = {
+    mode: 'development',
+    entry: './src/app.ts',
+    output: {
+        filename: 'bundle.js',
+        path: path.resolve(__dirname, 'dist'),
+        // public: 'dist',
+    },
+    devServer: {
+        publicPath: '/dist/',
+    },
+    devtool: 'inline-source-map',
+    module: {
+        rules: [
+            {
+                test: /\.ts$/,
+                use: 'ts-loader',
+                exclude: /node_modules/,
+            },
+        ],
+    },
+    resolve: {
+        extensions: ['.ts', '.js'],
+    },
+};
